### 2018-10-22 #1039 @whedon
[PRE REVIEW]: COSplay: Contrast Optimized Stimulation Player
**Submitting author:** @TheChymera (<a href="http://orcid.org/0000-0001-7037-2449">Horea-Ioan Ioanas</a>)
**Repository:** <a href="https://github.com/IBT-FMI/COSplay/" target ="_blank">https://github.com/IBT-FMI/COSplay/</a>
**Version:** 1.1.1
**Reviewers:** damaggu
