### 2018-10-24 1b3546c
Added link for COSplayer assembly
