### 2018-11-22 #1039 @labarba
@TheChymera : I had a quick look at your submission, and I'm a little concerned about how a reviewer might approach this. The software is meant to be used with a device, the **COSplayer**, which is custom-built open-source hardware using a PyBoard. Do you provide any way for the software to be tested _without_ this device? It would be useful, for example, to have synthetic data and a toy workflow that could be executed to demo the functionality, without the hardware. Is that possible?
