### 2018-11-25 #1039 @TheChymera
ugh, not really, the software interacts with the hardware. At the very least you would need a pyboard, and then you could observe the LED pattern.

If even that is unavailable you would need a software emulator for the pyboard, which is very much nontrivial to produce and would add little value to the project (others have attempted to make such emulators and given up). Unless such an emulator is guaranteed to be perfect (which is impossible), errors appearing in this use case might as well be errors with the emulator.

What I could offer, if a pyboard truly is unavailable, is to mail a pyboard to the reviewer (which we would like to have back :) ).

We had hoped that an interested reviewer might actually attempt to reproduce the device as per our complete video guide [[1](https://www.youtube.com/watch?v=YakrW-PoAgM),[2](https://figshare.com/articles/A_Guide_to_Assembling_the_COSplayer_an_Open_Source_Device_for_Microsecond-Range_Stimulus_Delivery_with_broad_Application_in_Biomedical_Engineering_and_fMRI/7227626)]. The best part about this is that the reviewer would then get to keep and use the COSplayer, which is a really cool device!
