### 2018-12-03 #1039 @danielskatz
👋 @arokem - would you be willing to edit this submission for JOSS? (reposting)
