### 2018-12-28 #1039 @arokem
@TheChymera : so sorry for the delay here.

I was thinking of asking folks in the PsychoPy developer community to review this. Do you have any other ideas for anyone who might be interested in reviewing this?
