### 2018-12-28 #1039 @TheChymera
I agree that there might be some overlap in the focus of these projects. Psychopy is more specifically targeted to high-level human applications and complex stimulus generation, whereas COSplay is mainly targeted at animal applications and easy integration with animal imaging workflows.

As this community is smaller and more tightly knit, I am afraid I could only recommend reviewers who actually work with us. However there might be people in the broader Python neuroimaging community who could offer valuable input not biased by prior exposure to the project. Perhaps @chrisfilo , @satra , @oesteban , @effigies , or @salma1601 ? But sadly I cannot comment on their willingness to review.
