### 2019-01-04 #1039 @arokem
@satra: are you possibly available to review this paper for JOSS?
