### 2019-01-07 #1039 @arokem
@effigies : are you available to review this paper for JOSS?
