### 2019-01-07 #1039 @TheChymera
@damaggu could help out. He also has access to a Bruker scanner similar to ours. I think he has already crossed paths with one of the copies of the device floating around campus, but he is not a contributor.
