### 2019-01-14 #1039 @arokem
I think that reproducing the device is too tall an order for a reviewer on the paper. I think that checking that everything matches the requirements and offering feedback could be done even without direct access to such a device. I suggest that @effigies and @damaggu both conduct the review and that you use the repo issues as a public place to discuss any issues between you.

Thanks everyone!
