### 2019-01-14 #1171 @whedon
[REVIEW]: COSplay: Contrast Optimized Stimulation Player
**Submitting author:** @TheChymera (<a href="http://orcid.org/0000-0001-7037-2449">Horea-Ioan Ioanas</a>)
**Repository:** <a href="https://github.com/IBT-FMI/COSplay/" target ="_blank">https://github.com/IBT-FMI/COSplay/</a>
**Version:** 1.1.1

@damaggu & @effigies, please carry out your reviews in this issue by updating the checklist below. If you cannot edit the checklist please:
