### 2019-01-14 #1039 @arokem
I'll close this one. Please head over to #1171 for the review.
