### 2019-01-14 #1039 @arfon
Just a heads up @arokem - using `assign` clobbers existing reviewer assignments. You need to use `add` to add additional reviewers when some already exist. I've fixed up the review issue over in #1171.
