### 2019-02-08 #1171 @effigies
RE conflicts of interest, I've met @TheChymera in person within the last 4 years (Scipy 2017), and we've collaborated (mostly in the form of PR review) separately on [Nipype](https://github.com/nipy/nipype/).

Also, this week piled up on me, and there's no chance I'll have a proper review today. I will aim for next Wednesday.
