### 2019-02-10 #1171 @arokem
Hi @effigies: the in-person meeting does not constitute a COI, in my opinion.

I am not sure whether the software collaboration constitutes COI, though. Let's check with the EIC. @arfon : the author of this paper and the reviewer have collaborated in code review on another software package (nipype, mentioned above). Does this constitute a COI that should disqualify @effigies from reviewing this paper? It is "work that is funded or published" only in the sense that that software project receives funding and is publicly available, but I believe that the two have not been paid from the same source of funding or been co-authors on any publication resulting from that software (but may in the future be). Should I find another reviewer for this?
