### 2019-02-11 #1171 @danielskatz
> @arokem, @danielskatz Sorry, I'll make some time this week as well. Regarding COI: I'm actually collaborating with @TheChymera on other projects, but haven't been involved in the development of COSplay. Does this constitute a problem? Regards

Yes, @arokem should try to find another reviewer in this case.
