### 2019-02-12 b46a50a
DOC: Fix typos in documentation, JOSS paper
