### 2019-02-12 #1171 @damaggu
@danielskatz Just wanted to point out that, @TheChymera is not the first author. I haven't interacted with the first outer, in case that makes a difference regarding COI.
