### 2019-02-12 #1171 @danielskatz
But he's the submitting author.  In my opinion, this is enough of a conflict with the submission that you shouldn't be asked to review it.
