### 2019-02-12 #1171 @danielskatz
JOSS allows recording of the COI and waiving it if that's the best case for the submission.  If @arokem thinks that how the review should proceed, it can.
