### 2019-02-15 #1171 @arokem
Yes. This submission is hard enough to review as it is, and the number of potential reviewers so small, that I think that we should go ahead with that in mind. Considering the review is open, it will be straightforward to see that the review is being done in earnest, despite COI.
