### 2019-02-18 #1171 @damaggu
@arokem @danielskatz I've found actually another potential reviewer on campus @tanselbaran, that would fulfil both the proximity and the COI requirement. Therefore I'd propose him as reviewer. Would that work?
