### 2019-02-19 #1171 @arokem
Thanks for finding an alternate reviewer! @tanselbaran: would you be willing to review this paper for JOSS? If you are, we can excuse @damaggu and move on with this with you as the reviewer. I'll have to figure out how to add you here.
