### 2019-03-27 #1171 @tanselbaran
By the way, I am a PhD student in the same university and department as the authors, but in different institutes and fields of research. I have never met any of them in person nor collaborated with them on any project. So there are no issues for me to review this work, right?
