### 2019-04-24 #1171 @tanselbaran
@arokem  Sorry for the delay. Yes, I have been looking into it further. I think this would be a useful package, but I currently have some concerns regarding the items that I didn't check off on the list. I am not familiar with the level of detail that is expected from the authors for the documentation and the example tests, but I couldn't really see any example usage or a simple test provided by the authors either in the repository or on the readthedocs page. There is a standalone test provided for the hardware and a very nice video provided on how to assemble the hardware, but the main scope of this publication is the software package, is that correct?

There is also quite a detailed documentation of individual functions that are part of the package. However, as someone who is completely unfamiliar with this package, I am quite lost on a higher level as to how I can get started with doing anything with this package, if that makes any sense. Maybe it would be more helpful for the future users if the authors could an example case of, for instance, how to generate a simple pulse train by using the COSplay package. That's why I refrain from checking off the items "Example usage" and "Automated Tests" at the moment. Since I couldn't get to test the function of the package, I am also refraining from checking off the "Functionality" item for now. I checked off the "Functionality documentation" item since the authors seems to have done a satisfactory job for documenting the functions included in the package.

As for the "Community guidelines", I couldn't really see any clear guidelines for future developers anywhere on the repository or the documentation page, but that might be just me not finding it. @effigies , did you see anything along those lines?

@arokem, please let me know if you have any questions or need further comments on any of the issues I have mentioned.
