### 2019-04-29 #1171 @arokem
@tanselbaran : thank you for your comments. These are all well within the scope of this review and should be addressed in a revision.
