### 2019-04-30 #1171 @effigies
Hi all, sorry about the delay as well. I've left the Functionality section of the review to @tanselbaran as discussed above. My concerns remain mostly with the documentation. In addition to the below, the documentation does not appear to be hosted, and I built it myself. I don't know if web-accessible documentation is a requirement for JOSS, so I'm just noting that.

> ### Documentation
>
>  **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?

This is present in the paper but not the documentation.

> **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

I see that dependencies are not listed in `setup.py`. Is there a principled reason not to do this? Also, as a Python package, people may expect it to be pip installable from PyPI. (I did install with `pip install .`, but being able to say `pip install COSplay==1.1.1` would be easier.)

> **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

I didn't see anything that really looked like this.

> **Automated tests:** Are there automated tests or manual steps described so that the function of the software can be verified?

There are tests of the client/server connection. The documentation has what appears to be a reasonable level of detail, but without a board, any gaps will not be evident to me. I'm checking this, with the understanding that the @tanselbaran's functionality checks will verify these.

> **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

These seem to be missing.
