### 2019-04-30 #1171 @TheChymera
@tanselbaran @effigies thank you for your comments. Could you please open issues ont he package so that I know exactly what I should add, and we can keep better track of when this has been resolved?

Regarding hosted docs, I do not thing this is required by JOSS, but they do exist:

https://cosplay.readthedocs.io/en/latest/?badge=latest
