### 2019-04-30 ab7279c
Updated dependency instructions
