### 2019-04-30 #1171 @TheChymera
Regarding the other issues of contributing guidelines and dependencies, these have now been resolved. If there is anything else left, please create an issue.
