### 2019-05-01 IBT-FMI/COSplay#37 @effigies
JOSS Documentation Review
From my JOSS review (https://github.com/openjournals/joss-reviews/issues/1171#issuecomment-487947018):

* The statement of need is present in the JOSS paper, but not in the documentation:

https://github.com/IBT-FMI/COSplay/blob/ab7279cd3ba7feaf4be199841cfb0fa0316ca4f3/paper.md#L36-L37

* Example usage of the software is missing

My other comments have been addressed.
