### 2019-05-01 #1171 @effigies
@arokem My two outstanding issues, corresponding to **A statement of need** and **Example usage**: https://github.com/IBT-FMI/COSplay/issues/37

For **Functionality**, **Performance** and **Automated tests**, I'm deferring to @tanselbaran's assessment of these and leaving them unchecked until he checks them off.
