### 2019-05-06 #1171 @arokem
@tanselbaran : where do you stand on testing the functionality of the software?
