### 2019-05-06 IBT-FMI/COSplay#38 @tanselbaran
Example usage and automated tests missing
See JOSS review criteria and my comment [https://github.com/openjournals/joss-reviews/issues/1171#issuecomment-486387960](url) for details
