### 2019-05-06 #1171 @tanselbaran
I was on vacation last week, but I can take a look this week. I opened an issue for the missing example usage and tests--I think I would need to have those available to test functionality.
