### 2019-05-14 2e6269b
Improved documentation as part of JOSS review
