### 2019-05-14 IBT-FMI/COSplay#38 @TheChymera
Dear @tanselbaran , thank you for taking the time to review our project. Does the new [README section](https://github.com/IBT-FMI/COSplay#minimal-usage-example) address your open concerns?
