### 2019-06-21 #1171 @tanselbaran
Dear all,

My apologies for the long delay. I just got around to testing the hardware, as things have been quite busy with my project.

After following the steps that are provided by the authors in the documentation of the repository, I verify that the device performs the test stimulus from all output channels with the specified frequencies and pulse widths. Thus, I am checking the functionality, performance and automated test items.

Two minor improvements to be made in the automated test description:
-The actual folder name in the device is named "sequence_library" rather than "sequences" as mentioned in the readme. Fixing that would remove any confusions from the user side.
-It wasn't very clear to me what the "amplitude" mentioned in the sequence.tsv files was. I am not sure what the 0.6 and 0.5 in rows 6 and 7 are referring to in the table.  According to SI units as mentioned by the readme file, they should be referring to values in Volts, but I believe that's not the case. As I could measure, the amplitude of the output pulse was 2 V for Channel 5 and 1.5 V for Channel 6.

Please let me know if there are any questions.
