### 2019-06-24 #1171 @arokem
Hello @TheChymera : looks like just a few small comments to address from @tanselbaran, in addition to previous comments from @effigies (described in https://github.com/openjournals/joss-reviews/issues/1171#issuecomment-488434380). Please let me know when you've had a chance to address these.
