### 2019-07-15 799f50d
Amplitude column clarification
