+++ README.rst
@@ -67,7 +67,7 @@ To execute a demo sequence on the COSplayer device, please follow the subsequent
 2. Connect the COSplayer device to the machine. At this point the blue, orange, and green LEDs should light up, according to the `LED legend <https://cosplay.readthedocs.io/en/latest/pyboard.html#led-pattern-legend>`_.
 3. Press the USR button twice. At this point the blue LED should light up, the red LED may also intermittently light up, according to whether data is being written on disk.
 4. After the device is recognised by the host computer, navigate to the `sequence_library` directory.
-5. Copy a file formatted according to the COSplay/BIDS standard into the directory. An example file is `provided in this repository <sequence.tsv>`_. All values are interpreted as SI base units, seconds and hertz, respectively.
+5. Copy a file formatted according to the COSplay/BIDS standard into the directory. An example file is `provided in this repository <sequence.tsv>`_. All values are interpreted as SI base units, seconds and hertz, respectively --- with the exception of the amplitude (`see more <https://cosplay.readthedocs.io/en/latest/pyboard.html#circuit>`_).
 6. After the file has copied and the red LED is no longer lit, safely remove the volume, and press the RST button.
 7. Click the USR button once to select the standalone operation mode. At this point the LED lighting should transition from blue, orange, and green to none.
 8. Click the USR button again to arm the device. At this point the orange LED should light up.
