### 2019-07-15 #1171 @TheChymera
@danielskatz @arokem sorry for the long delay, I too have been sidetracked by my thesis.

Addressing @effigies ' open concerns: https://github.com/IBT-FMI/COSplay/commit/ba9406744637701d1d586a28551e653960d4041b

Addressing @tanselbaran 's concerns: https://github.com/IBT-FMI/COSplay/commit/413c49149282bbbd39e5ea322fed492ac7aed724
Regarding the question about the amplitude: it is clarified [in the first two paragraphs, here](https://github.com/IBT-FMI/COSplay/blob/master/documentation/pyboard.rst#circuit). This section was now also linked to in the README overview: https://github.com/IBT-FMI/COSplay/commit/799f50d180375c8de17abac804e5dc866f60e328
