### 2019-07-15 #1171 @effigies
My concerns were satisfied, and I'm checking off the boxes that I deferred to @tanselbaran's judgment, as he has checked those.
