### 2019-07-16 #1171 @arokem
@tanselbaran : just to confirm -- are your comments all addressed?
