### 2019-07-16 #1171 @arokem
Great. I think that we ready to proceed with this.

@TheChymera : could you please create an archive of the software, as it is currently, or of a new release (e.g., on https://zenodo.org) and post the DOI here?
