### 2019-07-16 04b8d28
minor grammar change in paper
