### 2019-07-16 #1171 @danielskatz
@TheChymera  - can you accept https://github.com/IBT-FMI/COSplay/pull/39 please to fix an incorrect ; and slightly make the sentence that contains it more clear?
