### 2019-07-16 #1171 @TheChymera
@danielskatz merged. Do I need to update the archive now on zenodo?
